==> nbs/01_model_building/archs4/03_archs4_CLAMPfull.ipynb <==
# ARCHS4 model building with PLIER2

## Load libraries

if (!requireNamespace("PLIER2", quietly = TRUE)) {
    REPO_PATH <- "/home/msubirana/Documents/pivlab/PLIER2" 
    remotes::install_local(REPO_PATH, force = TRUE, dependencies = FALSE)
}

library(bigstatsr)
library(data.table)
library(dplyr)
library(rsvd)
library(glmnet)
library(Matrix)
library(knitr)
library(here)
library(PLIER2)
library(hdf5r)
library(biomaRt)

source(here("config.R"))

set.seed(123)

## Output directory

output_dir <- config$ARCHS4$DATASET_FOLDER
dir.create(output_dir, showWarnings = FALSE, recursive = TRUE)

archs4_svdRes <- readRDS(file.path(output_dir, "archs4_svdRes.rds"))

archs4_fbm_filt <- readRDS(file.path(output_dir, "archs4_fbm_filt.rds"))

archs4_baseRes <- readRDS(file.path(output_dir, "archs4_baseRes.rds"))

PLIER_K_archs4 <- readRDS(file.path(output_dir, "PLIER_K_archs4.rds"))

archs4_genes <- readRDS(file.path(output_dir, "archs4_genes.rds"))

sample_names <- readRDS(file.path(output_dir, "archs4_samples.rds"))

## Prepare pathway priors

# # run localy since server not have internet access

archs4_gmtList <- list(
  KEGG = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=KEGG_2021_Human"),
  BP = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=GO_Biological_Process_2025"),
  GTEx_Tissues = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=GTEx_Tissues_V8_2023")
)

# prefix each gene‐set name with its library to guarantee uniqueness
for(lib in names(archs4_gmtList)) {
  names(archs4_gmtList[[lib]]) <- paste0(lib, "_", names(archs4_gmtList[[lib]]))
}

archs4_pathMat <- gmtListToSparseMat(archs4_gmtList)
saveRDS(archs4_pathMat, file = file.path(output_dir, "archs4_pathMat.rds"))

archs4_pathMat <- readRDS(file.path(output_dir, "archs4_pathMat.rds"))
archs4_matched <- getMatchedPathwayMat(archs4_pathMat, archs4_genes)
archs4_chatObj <- getChat(archs4_matched)

saveRDS(archs4_pathMat, file = file.path(output_dir, "archs4_pathMat.rds"))
saveRDS(archs4_matched, file = file.path(output_dir, "archs4_matched.rds"))
saveRDS(archs4_chatObj, file = file.path(output_dir, "archs4_chatObj.rds"))

## PLIERfull

N_CORES <- config$ARCHS4$PLIER_PARAMS$N_CORES

archs4_fullRes <- PLIERfull(
  Y                 = archs4_fbm_filt,
  priorMat          = as.matrix(archs4_matched),
  svdres            = archs4_svdRes,
  plier.base.result = archs4_baseRes,
  Chat              = archs4_chatObj,
  k                 = PLIER_K_archs4,
  doCrossval        = TRUE,
  trace             = TRUE,
  max.U.updates     = 50,
  multiplier        = 3,
  ncores            = N_CORES
)

head(archs4_fullRes$summary) %>%  arrange(p_value)

# save summary as CSV
data.table::fwrite(as.data.frame(archs4_fullRes$summary),
                   file = file.path(output_dir, "archs4_fullRes_summary.csv"))

archs4_fullRes$summary %>%  arrange(`p-value`)

saveRDS(archs4_fullRes, file = file.path(output_dir, "archs4_PLIER2.rds"))

# Fix col, row names and summary

archs4_fullRes <- readRDS(file.path(output_dir, "archs4_PLIER2.rds"))

colnames(archs4_fullRes$B) <- sample_names

colnames(archs4_fullRes$Z) <- paste0('LV', seq_len(ncol(archs4_fullRes$Z)))

archs4_fullRes$summary <- archs4_fullRes$summary %>%
    dplyr::rename(LV = `LV index`) %>% 
    dplyr::mutate(LV = paste0('LV', LV))

saveRDS(archs4_fullRes, file = file.path(output_dir, "archs4_PLIER2.rds"))

archs4_baseRes <- readRDS(file.path(output_dir, "archs4_baseRes.rds"))

colnames(archs4_baseRes$B) <- sample_names
colnames(archs4_baseRes$Z) <- paste0('LV', seq_len(ncol(archs4_fullRes$Z)))
rownames(archs4_baseRes$Z) <- archs4_genes
rownames(archs4_baseRes$B) <- paste0('LV', seq_len(ncol(archs4_baseRes$Z)))

saveRDS(archs4_baseRes, file = file.path(output_dir, "archs4_baseRes.rds"))

archs4_fullRes$summary  %>%  arrange(`p-value`)
