==> nbs/01_model_building/archs4/06_archs4_CLAMPfull_M_multimodel.ipynb <==
# ARCHS4 model building with PLIER2

## Load libraries

if (!requireNamespace("PLIER2", quietly = TRUE)) {
    REPO_PATH <- "/home/msubirana/Documents/pivlab/PLIER2" 
    remotes::install_local(REPO_PATH, force = TRUE, dependencies = FALSE)
}

library(bigstatsr)
library(data.table)
library(dplyr)
library(rsvd)
library(glmnet)
library(Matrix)
library(knitr)
library(here)
library(PLIER2)
library(hdf5r)
library(biomaRt)

source(here("config.R"))

set.seed(123)

## Output directory

output_dir <- config$ARCHS4$DATASET_FOLDER
dir.create(output_dir, showWarnings = FALSE, recursive = TRUE)

output_dir_milton <- '/pividori_lab/projects/plier2/plier2_repo/output/archs4'

# meta
meta <- readRDS(file.path(output_dir_milton, "metadata_filtered.rds"))
n_genes_thin <- meta$n_genes_thin
n_samples <- meta$n_samples

# fbm
fbm_file  <- file.path(output_dir_milton, "fbm")
output_file <- paste0(fbm_file, "_filtered")

archs4_fbm_filt <- FBM(
  nrow        = n_genes_thin,
  ncol        = n_samples,
  backingfile = output_file ,
  create_bk   = FALSE,
)

# svd
archs4_svdRes <- readRDS(file.path(output_dir_milton, "svd.rds"))
archs4_genes <- meta$gene_symbols_thin

archs4_baseRes <- readRDS(file.path(output_dir, "archs4_baseRes.rds"))

PLIER_K_archs4 <- readRDS(file.path(output_dir, "PLIER_K_archs4.rds"))

sample_names <- readRDS(file.path(output_dir, "samples_archs4.rds"))

## Prepare pathway priors

# # run localy since server not have internet access
# KEGG_list <- list(
#   KEGG = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=KEGG_2021_Human")
# )

# for(lib in names(KEGG_list)) {
#   names(KEGG_list[[lib]]) <- paste0(lib, "_", names(KEGG_list[[lib]]))
# }

# KEGG_pathMat <- gmtListToSparseMat(KEGG_list)
# saveRDS(KEGG_pathMat, file = file.path(output_dir, "KEGG_pathMat.rds"))

# BP_list <- list(
#   BP = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=GO_Biological_Process_2025")
# )

# for(lib in names(BP_list)) {
#   names(BP_list[[lib]]) <- paste0(lib, "_", names(BP_list[[lib]]))
# }

# BP_pathMat <- gmtListToSparseMat(BP_list)
# saveRDS(BP_pathMat, file = file.path(output_dir, "BP_pathMat.rds"))

# GTEx_Tissues_list <- list(
#   GTEx_Tissues = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=GTEx_Tissues_V8_2023")
# )

# for(lib in names(GTEx_Tissues_list)) {
#   names(GTEx_Tissues_list[[lib]]) <- paste0(lib, "_", names(GTEx_Tissues_list[[lib]]))
# }

# GTEx_Tissues_pathMat <- gmtListToSparseMat(GTEx_Tissues_list)
# saveRDS(GTEx_Tissues_pathMat, file = file.path(output_dir, "GTEx_Tissues_pathMat.rds"))


# CellMarker_list <- list(
#   CellMarker = getGMT("https://maayanlab.cloud/Enrichr/geneSetLibrary?mode=text&libraryName=CellMarker_2024")
# )

# for(lib in names(CellMarker_list)) {
#   names(CellMarker_list[[lib]]) <- paste0(lib, "_", names(CellMarker_list[[lib]]))
# }

# CellMarker_pathMat <- gmtListToSparseMat(CellMarker_list)
# saveRDS(CellMarker_pathMat, file = file.path(output_dir, "CellMarker_pathMat.rds"))

## PLIERfull KEGG

# KEGG_pathMat <- readRDS(file.path(output_dir, "KEGG_pathMat.rds"))
# KEGG_matched <- getMatchedPathwayMat(KEGG_pathMat, archs4_genes)
# KEGG_chatObj <- getChat(KEGG_matched)

# N_CORES <- config$ARCHS4$PLIER_PARAMS$N_CORES

# archs4_fullRes <- PLIERfull(
#   Y                 = archs4_fbm_filt,
#   priorMat          = as.matrix(KEGG_matched),
#   svdres            = archs4_svdRes,
#   plier.base.result = archs4_baseRes,
#   Chat              = KEGG_chatObj,
#   k                 = PLIER_K_archs4,
#   doCrossval        = TRUE,
#   trace             = TRUE,
#   max.U.updates     = 50,
#   multiplier        = 3,
#   ncores            = N_CORES
# )

# # Fix col, row names and summary
# colnames(archs4_fullRes$B) <- sample_names

# colnames(archs4_fullRes$Z) <- paste0('LV', seq_len(ncol(archs4_fullRes$Z)))

# archs4_fullRes$summary <- archs4_fullRes$summary %>%
#     dplyr::rename(LV = `LV index`) %>% 
#     dplyr::mutate(LV = paste0('LV', LV))

# saveRDS(archs4_fullRes, file = file.path(output_dir, "archs4_PLIER2_KEGG.rds"))

## PLIERfull BP

# BP_pathMat <- readRDS(file.path(output_dir, "BP_pathMat.rds"))
# BP_matched <- getMatchedPathwayMat(BP_pathMat, archs4_genes)
# BP_chatObj <- getChat(BP_matched)

# N_CORES <- config$ARCHS4$PLIER_PARAMS$N_CORES

# archs4_fullRes <- PLIERfull(
#   Y                 = archs4_fbm_filt,
#   priorMat          = as.matrix(BP_matched),
#   svdres            = archs4_svdRes,
#   plier.base.result = archs4_baseRes,
#   Chat              = BP_chatObj,
#   k                 = PLIER_K_archs4,
#   doCrossval        = TRUE,
#   trace             = TRUE,
#   max.U.updates     = 50,
#   multiplier        = 3,
#   ncores            = N_CORES
# )

# # Fix col, row names and summary
# colnames(archs4_fullRes$B) <- sample_names

# colnames(archs4_fullRes$Z) <- paste0('LV', seq_len(ncol(archs4_fullRes$Z)))

# archs4_fullRes$summary <- archs4_fullRes$summary %>%
#     dplyr::rename(LV = `LV index`) %>% 
#     dplyr::mutate(LV = paste0('LV', LV))

# saveRDS(archs4_fullRes, file = file.path(output_dir, "archs4_PLIER2_BP.rds"))

## PLIERfull GTEx

# GTEx_Tissues_pathMat <- readRDS(file.path(output_dir, "GTEx_Tissues_pathMat.rds"))
# GTEx_Tissues_matched <- getMatchedPathwayMat(GTEx_Tissues_pathMat, archs4_genes)
# GTEx_Tissues_chatObj <- getChat(GTEx_Tissues_matched)

# N_CORES <- config$ARCHS4$PLIER_PARAMS$N_CORES

# archs4_fullRes <- PLIERfull(
#   Y                 = archs4_fbm_filt,
#   priorMat          = as.matrix(GTEx_Tissues_matched),
#   svdres            = archs4_svdRes,
#   plier.base.result = archs4_baseRes,
#   Chat              = GTEx_Tissues_chatObj,
#   k                 = PLIER_K_archs4,
#   doCrossval        = TRUE,
#   trace             = TRUE,
#   max.U.updates     = 50,
#   multiplier        = 3,
#   ncores            = N_CORES
# )

# # Fix col, row names and summary
# colnames(archs4_fullRes$B) <- sample_names

# colnames(archs4_fullRes$Z) <- paste0('LV', seq_len(ncol(archs4_fullRes$Z)))

# archs4_fullRes$summary <- archs4_fullRes$summary %>%
#     dplyr::rename(LV = `LV index`) %>% 
#     dplyr::mutate(LV = paste0('LV', LV))

# saveRDS(archs4_fullRes, file = file.path(output_dir, "archs4_PLIER2_GTEx.rds"))

## PLIERfull CellMarker

CellMarker_pathMat <- readRDS(file.path(output_dir, "CellMarker_pathMat.rds"))
CellMarker_matched <- getMatchedPathwayMat(CellMarker_pathMat, archs4_genes)
CellMarker_chatObj <- getChat(CellMarker_matched)

N_CORES <- config$ARCHS4$PLIER_PARAMS$N_CORES

archs4_fullRes <- PLIERfull(
  Y                 = archs4_fbm_filt,
  priorMat          = as.matrix(CellMarker_matched),
  svdres            = archs4_svdRes,
  plier.base.result = archs4_baseRes,
  Chat              = CellMarker_chatObj,
  k                 = PLIER_K_archs4,
  doCrossval        = TRUE,
  trace             = TRUE,
  max.U.updates     = 50,
  multiplier        = 3,
  ncores            = N_CORES
)

# Fix col, row names and summary
colnames(archs4_fullRes$B) <- sample_names

colnames(archs4_fullRes$Z) <- paste0('LV', seq_len(ncol(archs4_fullRes$Z)))

archs4_fullRes$summary <- archs4_fullRes$summary %>%
    dplyr::rename(LV = `LV index`) %>% 
    dplyr::mutate(LV = paste0('LV', LV))

saveRDS(archs4_fullRes, file = file.path(output_dir, "archs4_PLIER2_cellMarker.rds"))
